ajustes

+++ 2024/J_Cardozo/src/lib/Giroscopio/L3G4200D.h
@@ -10,7 +10,7 @@
 
 class L3G4200D {
 private:
-    float x, y, z;
+    int16_t x, y, z;
     uint8_t odr; //Taxa de Dados de Saída
     uint8_t scale; //Faixa de medicao
     uint8_t getODR(uint16_t odr);
